Section 3/src/src/requirements.txt
numpy>=1.17
torch>=1.4
pydicom>=1.4
nibabel>=3.0
matplotlib>=3.1
scipy>=1.4
Pillow>=7.0
